Sub router on dashboard page

+++ interface/src/pages/dashboard/index.tsx
@@ -1,15 +1,18 @@
 import { useContext, useEffect } from "react";
-import { useNavigate } from "react-router";
+import { Route, Routes, useNavigate } from "react-router";
 import { Navbar } from "../../components/nav";
-import { AuthContext } from "../../context/AuthContext";
+import { Auth, AuthContext } from "../../context/AuthContext";
+import { DashboardHome } from "./home";
+import { ProjectsPage } from "./projects";
 import { DashboardStyle } from "./style";
+import { TasksPage } from "./tasks";
 
 export function DashboardPage() {
     const auth = useContext(AuthContext);
     const navigate = useNavigate();
-
+    const storage: Auth = JSON.parse(localStorage.getItem("auth"));
     useEffect(() => {
-        if(auth.auth.token === "") {
+        if(auth.auth.token === "" && storage === null) {
             auth.setAuth({id: "", token: "", username: ""});
             localStorage.clear();
             navigate("/");
@@ -20,7 +23,11 @@ export function DashboardPage() {
         <DashboardStyle>
             <Navbar/>
             <main>
-                <h1>Welcome, User</h1>
+                <Routes>
+                    <Route path="/" element={<DashboardHome username={auth.auth.username || storage.username}/>}/>
+                    <Route path="/:projectId/tasks" element={<TasksPage/>}/>
+                    <Route path="/projects" element={<ProjectsPage/>}/>
+                </Routes>
             </main>
         </DashboardStyle>
     )
